requirements.txt
python-dotenv
google-generativeai
faster-whisper
gTTS
python-telegram-bot==20.6
